Add files via upload
A string or character array represents a mobile no. and an array containing five mobile numbers is provided. Verify these five mobile numbers and identify valid and invalid mobile numbers.
Any mobile number which have more or less than 10 digits is considered as invalid. Write a code which will generate a report on the count of total number of valid mobile numbers.
A valid mobile no. 'M' should be of length 10 with all numeric digits.